refactor: usePostDetail 훅 분리
